refactor repository to package
